feat(eslint-plugin): [no-unnecessary-condition] add rule
Adds a no-unnecessary-condition, a rule that uses type information to determine if a condition is necessary. Any condition that's indicated by types to always be truthy or falsy is considered to be unnecessary. Essentially a stronger version of no-constant-condition from core eslint.

For example:

```ts
const items = ['foo', 'bar'];
if(items) { } // Error: items is always truthy
```