chore: use discriminated unions for member name types
property names are either `computed: false` with `key: Identifier | StringLiteral | NumberLiteral`, or they are `computed: true` with `key: Expression`.
the previous typings took the simple approach of just having a single type, but it made things a bit more cumbersome.

This change creates two types for each, using the `computed` value as the key for a discriminated union.
This means that if you check `node.computed === true`, then TS will narrow the `key` type appropriately.

I also noticed a minor bug in the `TSEnumMember` handling, as the types didn't previously support the fact that it's syntactically valid to use an expression (note it's semantically invalid - TS error 1164).
It's also semantically and syntactically valid to do something like `enum Foo { ['key'] }`, which really should be handled differently to `enum Foo { key }`, even if they mean the same thing.
So I also added a `computed` prop to `TSEnumMember`, and gave it the same discriminated union treatment.