feat(eslint-plugin): [no-unsafe-array-reduce] Add rule
Add rule for protecting against unsafe array reduce calls.

The type of {} is {}. Any object is assignable to {}. {} is assignable
to any indexed type (`{[key in string]: whatever}`)

A reduce call with an empty object initializer and no type signature,
will infer the {} type for the accumulator and result of the reduce
expression. Since anything is assignable to {}, this means the reduce
function is essentially unchecked. The result of the expression can then
also be assigned to an incompatible type without raising any errors.

This rule warns if a reduce call takes an empty object as the initial
value and has no type signatures.