chore: separate linting checks to their own step
If linting/formatting/spelling fails, it should not block running the tests.
If they do block, they can increase the iteration time for contributors, because they will have to fix them before they can see the status of their tests.

By not blocking, a contributor will be able to submit a PR, and come back ~5min later knowing they will be able to see if their code passes linting _and_ tests.

This should also knock ~40-60s off of the perceived runtime, as it parallelises some of the work.