fix crash in unbound-method due to missing Intl
When Node.js is compiled with the "--with-intl=none" option ( https://nodejs.org/api/intl.html#intl_options_for_building_node_js ) it will not have the Intl global. As a result, any attempt to run eslint would result in an error like:

  TypeError: Failed to load plugin '@typescript-eslint' declared in '.eslintrc.js': Cannot convert undefined or null to object
  Referenced from: /home/dev/web/.eslintrc.js
      at Function.getOwnPropertyNames (<anonymous>)
      at /home/dev/web/node_modules/@typescript-eslint/eslint-plugin/dist/rules/unbound-method.js:91:19

We can silently ignore missing globals as a workaround.