docs: add link to v4 release notes to the changelogs
Some users use CHANGELOG.md as the source of truth for releases and don't check the github releases page.
We should probably delete these files and just have everything in the GH releases page.

This adds a link to the  v4 releases page and adds a disclaimer to the top of the file pointing at the releases page.