fix(prefer-ts-expect-error): change fix to suggestion
There are use-cases for ts-ignore - somewhat rare, but legitimate. For example, usages of recent typescript features like variadic tuple types, with graceful downgrading:

```typescript
// @ts-ignore
type Push_ts4<A extends unknown[], B> = [...A, B];
type VariadicTuplesPrefixesSupported = Push_ts4<[1, 2], 3> extends { length: 3 } ? "yes" : "no";
type Push<A extends unknown[], B> = VariadicTuplesPrefixesSupported extends "yes"
    ? Push_ts4<A, B>
    : Array<A[number] | B>;
```

This will be a typescript error for users on typescript 3.9 or below, but will compile fine for typescript 4.0 and above. If we autofix to ts-expect-error, a developer might not notice it changing and have everything compile fine, only for another user to hit an "unused ts-expect-error" directive.

This change swaps the `fix` for a `suggest` so that the developer has to explicitly choose whether to apply the fix or determine whether it's a legitimate use of `ts-ignore`.