chore(eslint-plugin): [naming-convention] refactor rule to split it up
The rule file was up to 1700 LOC.
It was a pain in the butt to scroll around it and find pieces.
I'm pretty sure ESLint / TypeScript gets a bit choked up on a file that large as well (I've been running into all sorts of slowness with it).

So simply isolate each part into a module to better separate things and reduce the total LOC per file.

Also moved some stuff around in the docs to try and focus each section, and add an FAQ section for the misc stuff