fix(typescript-estree): add default value for `parserOptions.projectFolderIgnoreList` and deduplicate resolved projects
In #2418 I introduced a regression - I forgot to add in the default value for `projectFolderIgnoreList`.
This means that globs have been matching `node_modules` since the v4.0 release! Oops :(

This PR fixes that.
It also hoists the tsconfig path canonicalisation up so that we can deduplicate early.
Previously if you provided a config like `projects: ['./tsconfig.json', './**/tsconfig.json']`, then we would resolve this to `./tsconfig.json` and `tsconfig.json`, then later canonicalise them.
This meant we'd check that same tsconfig twice!
By hoisting the canonicalisation, we can deduplicate this early.

Fixes #2814