Add fix JSON example to no-undef in FAQ.md
I had the same questions as the user in #2799 upon upgrading to v.4, and wasn't sure what the proposed solution to the problem here was.  I re-ordered the prose so that it gives the recommended solution first, then a code example of how to safely turn off no-undef in a mixed JS/TS environment, and rephrased the "edit the globals" recommendation as more of a "last-resort" type of solution.