docs(ast-spec): add script to generates code strings for the doc files
I'm not sure if we will want to actually use this or not.
This was just an idea I had to allow us to embed the relevant code samples directly in a `.md` file saving consumers from having to switch between the .ts file and the .md file.

We could ofc just embed the file itself, but that's not as elegant as there's the extra cruft of `export` and  `import`s.
May end up just ditching this... but it was worth putting it up for posterity.

As an example - this script generates the following output for `ClassProperty.ts`:

```ts
interface ClassPropertyComputedName extends BaseNode {
  type: 'ClassProperty';
  key: Expression;
  computed: true;
  value: Expression | null;
  static: boolean;
  declare: boolean;
  readonly?: boolean;
  decorators?: Decorator[];
  accessibility?: "private" | "protected" | "public";
  optional?: boolean;
  definite?: boolean;
  typeAnnotation?: TSTypeAnnotation;
}

interface ClassPropertyNonComputedName extends BaseNode {
  type: 'ClassProperty';
  key: Identifier | NumberLiteral | StringLiteral;
  computed: false;
  value: Expression | null;
  static: boolean;
  declare: boolean;
  readonly?: boolean;
  decorators?: Decorator[];
  accessibility?: "private" | "protected" | "public";
  optional?: boolean;
  definite?: boolean;
  typeAnnotation?: TSTypeAnnotation;
}

type ClassProperty =
  | ClassPropertyComputedName
  | ClassPropertyNonComputedName;
```