docs: remove node_modules from .eslintignore
I found this in the documentation of ESLint (https://eslint.org/docs/user-guide/configuring/ignoring-code)

In addition to any patterns in the .eslintignore file, ESLint always follows a couple of implicit ignore rules even if the --no-ignore flag is passed. The implicit rules are as follows:
- node_modules/ is ignored.
- dot-files (except for .eslintrc.*), as well as dot-folders and their contents, are ignored.

Maybe it's confusing to the person reading the documentation?