fix(typescript-estree): stub out `ts.SatisfiesExpression` on old TS versions <!-- 👋 Hi, thanks for sending a PR to typescript-eslint! 💖 Please fill out all fields below and make sure each item is true and [x] checked. Otherwise we may not be able to review your PR. -->
## PR Checklist

- [x] Steps in [CONTRIBUTING.md](https://github.com/typescript-eslint/typescript-eslint/blob/main/CONTRIBUTING.md) were taken

## Overview

<!-- Description of what is changed and how the code change does that. -->

Another thing we need to add to the "new syntax" contributor guide.
This union is part of the API surface for the package, so we need to ensure we don't break the build on old TS versions by "stubbing" out the new AST node types.

Also added version comments so it's easy to tell when we can remove them as we bump the minimum TS version