docs: Fix ignorePrimitives examples in prefer-nullish-coalescing docs
The examples for the `ignorePrimitives` option were flipped. Setting `ignorePrimitives` to true means that any logical OR expression with a primitive operand is ignored by the lint rule.

I also made two other edits to this docs page. The first is to change "you can safely replace" to "you may consider replacing" in the lint rule's description. The reason for this change is that this lint rule suggests changes that alter the code's behavior in subtle cases where `x || y !== x ?? y`.

There is also a small change to reword the tense in one of the sentences to help it read more smoothly.