fix(rule-tester): fix `parser` fallback logic
The docs at https://typescript-eslint.io/packages/rule-tester/#type-aware-testing
show that for type-aware rule testing `parserOptions` needs to be
specified. However, like shown in the docs, when `parser` is omitted
then an error: "The first argument to require.resolve must be a string.
Received null or undefined" is thrown due to a bug in `parser` fallback
behavior.