fix(utils): allow an array for the values of SharedConfig
According to the docs for ESLint an array should be allowed here, and
the type used by ESLint in their `ESLint.Plugin` interface is:

    Record<string, Linter.LegacyConfig | Linter.Config | Linter.Config[]> | undefined

So we should also allow an array. The helper function `tseslint.config`
returns an array so without this you're not allowed to use a config
returned by this function as a value in this record.

Fixes #10213