chore: add test-coverage script
Adds a `test-coverage` script to each package.

With this change, each package has these two scripts:

- `test` - run tests with coverage disabled
- `test-coverage` - run tests with coverage enabled (_if_ the package
  wants to support coverage)

Before this change, `test` behaved like `test-coverage` (i.e. packages
had coverage enabled in it). In CI, we then forced coverage off in some
situations by passing a flag: `npm run test -- --coverage=false`.

This resulted in commands like `jest --coverage --coverage=false`.

While jest is happy with this, vitest is not. Given we are likely to
move to vitest, it is worth fixing this ahead of time.