chore(eslint-plugin): no-confusing-void-expression `getConstraintInfo`
Moves to using `getConstraintInfo` internally in the
`no-confusing-void-expression` rule.

Also fixes a bug this uncovered, where `AwaitExpression` nodes were
actually untested and not supported in the type constraint checks.

This is because we passed `node` (the `AwaitExpression` itself), rather
than the thing being awaited (`node.argument`). By passing the
`argument`, we check its type instead.