chore: update all test scripts to remove coverage flag (#10582)
* chore: add test-coverage script

Adds a `test-coverage` script to each package.

With this change, each package has these two scripts:

- `test` - run tests with coverage disabled
- `test-coverage` - run tests with coverage enabled (_if_ the package
  wants to support coverage)

Before this change, `test` behaved like `test-coverage` (i.e. packages
had coverage enabled in it). In CI, we then forced coverage off in some
situations by passing a flag: `npm run test -- --coverage=false`.

This resulted in commands like `jest --coverage --coverage=false`.

While jest is happy with this, vitest is not. Given we are likely to
move to vitest, it is worth fixing this ahead of time.

* chore(ast-spec): enable test coverage

* chore(scope-manager): add jest as dependency

* chore: drop coverage script

* chore: disable coverage in jest base config

* review changes

---------

Co-authored-by: Kirk Waiblinger <53019676+kirkwaiblinger@users.noreply.github.com>